when using content search, need to specify return attributes in the search string
